[SPOT-301][REFACTOR] OAuth 전략 패키지 구조 정리 및 오타 수정
