+++ src/main/java/com/example/spot/auth/application/refactor/strategy/OAuthStrategy.java
@@ -1,9 +1,10 @@
-package com.example.spot.auth.application.refactor;
+package com.example.spot.auth.application.refactor.strategy;
 
 import com.example.spot.member.domain.Member;
 import com.example.spot.member.domain.enums.LoginType;
 
 public interface OAuthStrategy {
+    
     LoginType getType();
 
     String getOauthRedirectURL();
