+++ src/main/java/com/example/spot/auth/application/refactor/strategy/provider/NaverOAuthStrategy.java
@@ -1,7 +1,7 @@
-package com.example.spot.auth.application.refactor.impl.straegy;
+package com.example.spot.auth.application.refactor.strategy.provider;
 
-import com.example.spot.auth.application.refactor.OAuthStrategy;
-import com.example.spot.auth.application.refactor.impl.oauth.NaverOauth;
+import com.example.spot.auth.application.refactor.strategy.OAuthStrategy;
+import com.example.spot.auth.infrastructure.oauth.NaverOauth;
 import com.example.spot.auth.presentation.dto.oauth.naver.NaverOAuthTokenDTO;
 import com.example.spot.auth.presentation.dto.oauth.naver.NaverUser;
 import com.example.spot.member.domain.Member;
